requirements.dev.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-login
flask-wtf
WTForms
# load the .env file
python-dotenv
# For production use
waitress
# Testing
bcrypt
faker 